+++ BookingSystem/BookingSystem.API/Data/SeedData.cs
@@ -1,3 +1,4 @@
+using BookingSystem.Domain.Constants;
 using Microsoft.AspNetCore.Identity;
 
 namespace BookingSystem.API.Data;
@@ -9,7 +10,7 @@ public class SeedData
         using var scope = services.CreateScope();
         var roleManager = scope.ServiceProvider.GetRequiredService<RoleManager<IdentityRole<Guid>>>();
 
-        foreach (var role in new[] { "Client", "Host", "Admin" })
+        foreach (var role in Roles.All)
         {
             if (!await roleManager.RoleExistsAsync(role))
                 await roleManager.CreateAsync(new IdentityRole<Guid>(role));
