+++ BookingSystem/BookingSystem.API/appsettings.Development.json
@@ -4,5 +4,15 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost;Database=BookingSystemDb;Trusted_Connection=True;MultipleActiveResultSets=true;Encrypt=False"
+  },
+
+  "JwtSettings": {
+    "Secret": "YOUR_SUPER_SECRET_KEY_FOR_DEV",
+    "Issuer": "BookingSystemAPI",
+    "Audience": "BookingSystemClient",
+    "ExpiresInMinutes": 60
   }
 }
