+++ BookingSystem/BookingSystem.Application/DTOs/LoginDto.cs
@@ -1,3 +1,3 @@
-namespace BookingSystem.API.DTOs;
+namespace BookingSystem.Application.DTOs;
 
 public record LoginDto(string Email, string Password);
