+++ BookingSystem/BookingSystem.Application/DTOs/RegisterDto.cs
@@ -1,3 +1,3 @@
-namespace BookingSystem.API.DTOs;
+namespace BookingSystem.Application.DTOs;
 
 public record RegisterDto(string FirstName, string LastName, string Email, string Password, string Role);
