+++ BookingSystem/BookingSystem.Application/IAuthService.cs
@@ -1,0 +1,9 @@
+using BookingSystem.Application.DTOs;
+
+namespace BookingSystem.Application;
+
+public interface IAuthService
+{
+    Task<AuthResultDto> RegisterAsync(RegisterDto dto);
+    Task<AuthResultDto> LoginAsync(LoginDto dto);
+}
